refactor: replace custom cuDNN setup script with initialized sub-repository for sonitranslate engine
