+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "packageManager": "bun@1.3.11",
   "scripts": {
-    "setup:api": "uv sync && uv run python scripts/setup_cudnn.py",
+    "setup:api": "uv sync && uv run python scripts/setup.py",
     "dev:api": "uv run uvicorn main:app --app-dir backend --host 0.0.0.0 --port 3900 --reload",
     "dev:frontend": "bun run --cwd frontend dev",
     "dev:desktop": "bun run --cwd frontend desktop",
